Changed : some comments on parse v2 Changed : seperate file "base.go" for embedded OD.
